# Retain `RenderMeshInstance` and `MeshInputUniform` data from frame to frame.